Refactor and cleanup code: - move definition of common modules in a cff   (cfgs cannot import from cfgs) - modify the treeProducer so that specific ones can inherit   from the susyCore one and just add variables - rename the SUSYGenStatusThree.py so that it follows a more   consistent naming convention
